streamlit-app/requirements.txt
requests==2.31.0
streamlit==1.28.1
python-dotenv==1.0.0
streamlit-cookies-manager==0.2.0
